No point in having separate module.
